feat : message update 및 delete 구현
- update 에 파라미터를 dto 으로 변경
- 삭제시 모든 binaryContent 삭제

+++ src/main/java/com/sprint/discodeit/service/file/FileMessageService.java
@@ -1,13 +1,16 @@
 package com.sprint.discodeit.service.file;
 
 import com.sprint.discodeit.domain.dto.messageDto.MessageRequestDto;
+import com.sprint.discodeit.domain.dto.messageDto.MessageUpdateRequestDto;
 import com.sprint.discodeit.domain.entity.BinaryContent;
 import com.sprint.discodeit.domain.entity.Message;
 import com.sprint.discodeit.domain.mapper.MessageMapper;
+import com.sprint.discodeit.repository.file.BaseBinaryContentRepository;
 import com.sprint.discodeit.repository.file.FileMessageRepository;
 import com.sprint.discodeit.service.MessageServiceV1;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,6 +21,8 @@ public class FileMessageService implements MessageServiceV1 {
 
     private final FileMessageRepository fileMessageRepository;
     private final BinaryContentService binaryContentService;
+    private final BaseBinaryContentRepository baseBinaryContentRepository;
+
 
     @Override
     public Message create(MessageRequestDto messageRequestDto) {
@@ -31,25 +36,36 @@ public class FileMessageService implements MessageServiceV1 {
 
     @Override
     public Message find(UUID messageId) {
-        Message message = fileMessageRepository.findById(messageId).orElseThrow(() -> new NoSuchElementException(messageId + " 없는 회원 입니다"));;
+        Message message = fileMessageRepository.findById(messageId).orElseThrow(() -> new NoSuchElementException(messageId + " 없는 회원 입니다"));
         return message;
     }
 
     @Override
-    public List<Message> findAll(UUID channelId) {
-        return fileMessageRepository.findByChannelAndMessageContext(channelId)
+    public List<Message> findAllByChannelId(UUID channelId) {
+        List<Message> message = fileMessageRepository.findByChannelAndMessageContext(channelId);
+        if (message.isEmpty() || message.size() == 0) {
+          throw new IllegalArgumentException("유효하지 않는 채널 입니다.");
+        }
+        return message;
     }
 
     @Override
-    public Message update(UUID messageId, String newContent) {
-        Message message = fileMessageRepository.findById(messageId).orElseThrow(() -> new NoSuchElementException(messageId + " 없는 회원 입니다"));;
-        message.update(newContent);
+    public Message update(MessageUpdateRequestDto messageUpdateRequestDto) {
+        Message message = fileMessageRepository.findById(messageUpdateRequestDto.messageId()).orElseThrow(() -> new NoSuchElementException(messageUpdateRequestDto.messageId() + " 없는 회원 입니다"));
+        message.update(messageUpdateRequestDto.newContent());
         return message;
     }
 
     @Override
     public void delete(UUID messageId) {
-        fileMessageRepository.delete(messageId);
+        Optional<Message> message = fileMessageRepository.findById(messageId);
+        if(message.isPresent()){
+            for(UUID binaryContentsId : message.get().getAttachmentIds()){
+                baseBinaryContentRepository.delete(binaryContentsId);
+            }
+            fileMessageRepository.delete(messageId);
+        }else{
+            throw new IllegalArgumentException("없는 메세지 입니다.");
+        }
     }
-
 }
